commit bài tập

+++ case-study/furama/src/main/java/com/example/model/entity/AttachService.java
@@ -10,7 +10,7 @@ import java.util.Set;
 public class AttachService {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-private int attachServiceId;
+private Long attachServiceId;
 private String attachServiceName;
 private double attachServiceCost;
 private int attachServiceUnit;
@@ -22,7 +22,7 @@ private String attachServiceStatus;
     public AttachService() {
     }
 
-    public AttachService(int attachServiceId, String attachServiceName, double attachServiceCost, int attachServiceUnit, String attachServiceStatus, Set<ContractDetail> contractDetailSet) {
+    public AttachService(Long attachServiceId, String attachServiceName, double attachServiceCost, int attachServiceUnit, String attachServiceStatus, Set<ContractDetail> contractDetailSet) {
         this.attachServiceId = attachServiceId;
         this.attachServiceName = attachServiceName;
         this.attachServiceCost = attachServiceCost;
@@ -31,11 +31,11 @@ private String attachServiceStatus;
         this.contractDetailSet = contractDetailSet;
     }
 
-    public int getAttachServiceId() {
+    public Long getAttachServiceId() {
         return attachServiceId;
     }
 
-    public void setAttachServiceId(int attachServiceId) {
+    public void setAttachServiceId(Long attachServiceId) {
         this.attachServiceId = attachServiceId;
     }
